feat(ux): game status
